style: move navigation
- move navigation to own component
- move login button under input
